requirements.txt
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
idna==3.13
oauthlib==3.3.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
requests==2.33.1
requests-oauthlib==2.0.0
schedule==1.2.2
urllib3==2.6.3
